renamed formfile to userAddressForm, deleted modal file -- will recreate in sep PR
