+++ SweetTooth.ui/src/components/UserProfile.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Container, Row } from 'reactstrap';
-import LandingPageForm from './forms/LandingPageForm';
+import LandingPageForm from './forms/userAddressForm';
 // import { getAllAddressesByUserId } from '../helpers/data/userAddressData';
 
 export default function UserProfile({
